dlgm/hongli/project_02/requirements.txt
aim==3.27.0
aim-ui==3.27.0
aimrecords==0.0.7
aimrocks==0.5.2
aiofiles==24.1.0
alembic==1.14.0
annotated-types==0.7.0
anyio==4.8.0
base58==2.0.1
boto3==1.35.97
botocore==1.35.97
Brotli @ file:///croot/brotli-split_1736182456865/work
cachetools==5.5.0
certifi @ file:///croot/certifi_1734473278428/work/certifi
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
contrastive-learner==0.1.1
cryptography==44.0.0
cycler==0.12.1
decorator==5.1.1
docker-pycreds==0.4.0
dominate==2.9.1
einops==0.8.0
fastapi==0.115.6
filelock @ file:///work/perseverance-python-buildout/croot/filelock_1701733993137/work
fire==0.7.0
fonttools==4.55.3
fsspec @ file:///croot/fsspec_1736274672181/work
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.1.1
h11==0.14.0
h5py==3.12.1
idna @ file:///croot/idna_1714398848350/work
Jinja2 @ file:///croot/jinja2_1730902924303/work
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.8
kornia==0.8.0
kornia_rs==0.1.8
Mako==1.3.8
MarkupSafe @ file:///croot/markupsafe_1736366905094/work
matplotlib==3.10.0
mkl-service==2.4.0
mkl_fft @ file:///io/mkl313/mkl_fft_1730824109137/work
mkl_random @ file:///io/mkl313/mkl_random_1730823916628/work
mpmath @ file:///work/perseverance-python-buildout/croot/mpmath_1698864994882/work
networkx @ file:///croot/networkx_1720002482208/work
numpy @ file:///croot/numpy_and_numpy_base_1708638617955/work/dist/numpy-1.26.4-cp312-cp312-linux_x86_64.whl#sha256=1d700f51d8b4fa684d858c9e3b56b1656bc5c82b6b79ff08d4e3b491c430059f
opencv-python==4.10.0.84
packaging==24.2
pandas==2.2.3
pillow @ file:///croot/pillow_1734430590456/work
platformdirs==4.3.6
protobuf==5.29.3
psutil==6.1.1
py==1.11.0
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
pyparsing==3.2.1
PySocks @ file:///work/perseverance-python-buildout/croot/pysocks_1698845478203/work
python-dateutil==2.9.0.post0
pytorch-fid==0.3.0
pytz==2024.2
PyYAML @ file:///croot/pyyaml_1728657952215/work
requests @ file:///croot/requests_1730999120400/work
RestrictedPython==7.4
retry==0.9.2
s3transfer==0.10.4
scipy==1.15.1
sentry-sdk==2.19.2
setproctitle==1.3.4
setuptools==75.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.41.3
stylegan2_pytorch==1.9.0
sympy==1.13.1
termcolor==2.5.0
torch==2.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
triton==3.1.0
typing_extensions @ file:///croot/typing_extensions_1734714854207/work
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.34.0
vector_quantize_pytorch==0.1.0
visdom==0.2.4
wandb==0.19.2
websocket-client==1.8.0
websockets==14.1
wheel==0.44.0
